compiler-guided fix of type conversion warning for inventory::size(); impose const correctness while at this

+++ npctalk.cpp
@@ -1256,7 +1256,7 @@ void talk_function::give_equipment(game *g, npc *p)
  p->init_selling(giving, prices);
  int chosen = -1;
  if (giving.empty()) {
-  for (int i = 0; i < p->inv.size(); i++) {
+  for (size_t i = 0; i < p->inv.size(); i++) {
    int val = p->value(p->inv[i]) - p->inv[i].price() / 50;
    giving.push_back(i);
    prices.push_back(p->inv[i].price());
